Set pins for fans
In My setup, I'm using pin D5 from the servos to drive a Mosfet on Pin D5 that turns spins a 12V fan.
#define E0_AUTO_FAN_PIN 5

Also, in my setup, pin D4 drives a fan that cools the electronics box.

In the Configuration.h file we undefine the servo2_pin, and there is a commented out FAN1_PIN 5, this was because i was making tests manually managing the pin with the GCODE M106 P1 S255, to set the temperature, since I was getting errors from thermal runaways because of too much cooling.

I ended up disabling this when I realized the heating cartridge was faulty, but I leave it here like this for future reference.